requirements.txt
amqp==2.6.1
appdirs==1.4.4
billiard==3.6.4.0
black==21.7b0
cached-property==1.5.2
celery==4.4.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==2.2.4
django-tastypie==0.14.2
flake8==3.9.2
flake8-isort==3.0.1
idna==2.10
importlib-metadata==4.6.1
isort==4.3.21
kombu==4.6.11
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
prompt-toolkit==3.0.19
pycodestyle==2.7.0
pyflakes==2.3.1
pyRFC3339==1.1
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2021.1
regex==2021.8.3
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
testfixtures==6.18.0
toml==0.10.2
tomli==1.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
vine==1.3.0
wcwidth==0.2.5
zipp==3.5.0
